+++ cse/pipeline/Handler.py
@@ -1,10 +1,24 @@
-from cse.pipeline import Handler
+import abc
 
-class SimpleConsolePrintHandler(Handler):
-    def process(self, ctx, data):
-        print(str(data))
-        ctx.write(data)
+"""
+Abstract class representing an handler processing data which flows
+through the pipeline.
+"""
+class Handler(object, metaclass=abc.ABCMeta):
+    __name = ""
 
-class SimpleForwardHandler(Handler):
+    def __init__(self, name):
+        self.__name = name
+
+    def registeredAt(self, ctx):
+        pass
+
+    @abc.abstractmethod
     def process(self, ctx, data):
-        ctx.write(data)
+        raise NotImplementedError(
+            "Class %s doesn't implement process(), please do that yourself" %
+            (self.__class__.__name__)
+        )
+
+    def __str__(self):
+        return self.__name
